Change defaults for fidelity bond settings.
Closes #1247.
This consists of a change to the default value of
bondless_makers_allowance, an inclusion of the bond value exponent into
the config that the user can alter, and a change to the default of that
value from 2 (implicit by quadratic) to 1.3.
Also updates the fidelity bond documentation to account for those
changes.